Add the generated sources for the libmusicbrainz bindings.
Since these only ever have to be regenerated if the bindings generation
or musicbrainz api changes, I'm just including the prebuilt files in the
repository to simplify the build system.

Updating the sources if needed later is just a matter of manually
running 'make sources' in the musicbrainz4 directory.

References: gh-8